Refactoring, corrected position of app search box on other app store pages
