+++ frontend-web/webclient/app/Applications/Landing.tsx
@@ -24,6 +24,7 @@ import {ContextSwitcher} from "@/Project/ContextSwitcher";
 import ApplicationRow from "./ApplicationsRow";
 import { GradientWithPolygons } from "@/ui-components/GradientBackground";
 import {CSSVarCurrentSidebarStickyWidth} from "@/ui-components/Sidebar";
+import {AppSearchBox} from "./Search";
 
 export const ApiLike: ReducedApiInterface = {
     routingNamespace: "applications",
@@ -88,56 +89,6 @@ const LandingDivider = injectStyle("landing-divider", k => `
         transform: scaleX(-1);
     }
 `);
-
-  const AppSearchBoxClass = injectStyle("app-search-box", k => `
-    ${k} {
-        width: 300px;
-        position: relative;
-        margin-right: 15px;
-    }
-
-    ${k} input {
-        width: 100%;
-        border: 1px solid var(--midGray);
-        background: var(--white);
-        border-radius: 6px;
-        padding-left: 1.2em;
-        padding-right: 2.5rem;
-    }
-
-    ${k} button {
-        background: none;
-        border: 0;
-        padding: 0px 10px 1px 10px;
-        cursor: pointer;
-        position: absolute;
-        right: 0;
-        height: 2.4rem;
-    }
-`);
-
-export const AppSearchBox: React.FunctionComponent<{value?: string}> = props => {
-    const navigate = useNavigate();
-
-    return <div className={AppSearchBoxClass}>
-        <Flex justifyContent="space-evenly">
-            <Input
-                defaultValue={props.value}
-                placeholder="Search for applications..."
-                onKeyUp={e => {
-                    console.log(e);
-                    if (e.key === "Enter") {
-                        navigate(AppRoutes.apps.search((e as unknown as {target: {value: string}}).target.value));
-                    }
-                }}
-                autoFocus
-            />
-            <button>
-                <Icon name="search" size={20} color="darkGray" my="auto" />
-            </button>
-        </Flex>
-    </div>;
-}
 
 const ApplicationsLanding: React.FunctionComponent = () => {
     const [sections, fetchSections] = useCloudAPI<AppStoreSections>(
